requirements.txt

streamlit
langchain
langgraph
openai
tavily-python
faiss-cpu
python-dotenv
